refactor: separate out project, task and db items
