+++ api/main.py
@@ -1,0 +1,30 @@
+from fastapi import FastAPI
+from contextlib import asynccontextmanager
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+from api.db.db import create_db_and_tables
+from api.tasks.tasks import task_router
+from api.projects.projects import project_router
+
+@asynccontextmanager
+async def lifespan(app: FastAPI):
+    create_db_and_tables()
+    yield
+
+
+app = FastAPI(lifespan=lifespan)
+app.include_router(task_router)
+app.include_router(project_router)
+origins = [
+    "*",
+]
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=origins,
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
+
+
